refactor(spdx-utils): Split the large `SpdxExpressionTest`
Improve the overview before adding more tests.
